Fix circular dependency between auth and users routes
The server was crashing silently on startup due to a circular
dependency between `routes/auth.js` and `routes/users.js`.
`routes/users.js` was requiring `routes/auth.js` to get user data,
while `server.js` was requiring both.

This change refactors the in-memory user data into a separate
module, `data/users.js`, which is now required by both `auth.js` and
`users.js`. This breaks the circular dependency and allows the server
to start correctly.

A `.gitignore` file has also been added to exclude log files and
other unnecessary files from version control.
