+++ .env
@@ -1,16 +1,45 @@
-# Test Environment Configuration
+# Environment Configuration for Méliès Online Improved
+
+# Application Environment
 NODE_ENV=development
-PORT=3001
+PORT=3000
 
 # JWT Configuration
-JWT_SECRET=test-super-secret-jwt-key-for-development-only-min-32-chars
+JWT_SECRET=your-super-secret-jwt-key-change-in-production-min-32-chars
 JWT_EXPIRES_IN=15m
 REFRESH_TOKEN_EXPIRES_IN=7d
 
-# Vimeo API Configuration (mock values for testing)
-VIMEO_CLIENT_ID=test_client_id
-VIMEO_CLIENT_SECRET=test_client_secret
-VIMEO_ACCESS_TOKEN=test_access_token
+# Vimeo API Configuration
+VIMEO_CLIENT_ID=your_vimeo_client_id
+VIMEO_CLIENT_SECRET=your_vimeo_client_secret
+VIMEO_ACCESS_TOKEN=your_vimeo_access_token
+
+# Database Configuration (PostgreSQL)
+DATABASE_URL=postgresql://username:password@localhost:5432/melies_online
+DB_HOST=localhost
+DB_PORT=5432
+DB_NAME=melies_online
+DB_USER=melies_user
+DB_PASSWORD=secure_password
+DB_SSL=false
+
+# Redis Configuration (for sessions and caching)
+REDIS_URL=redis://localhost:6379
+REDIS_HOST=localhost
+REDIS_PORT=6379
+REDIS_PASSWORD=
+
+# Email Configuration (for notifications)
+SMTP_HOST=smtp.gmail.com
+SMTP_PORT=587
+SMTP_USER=your-email@gmail.com
+SMTP_PASS=your-app-password
+FROM_EMAIL=noreply@melies-online.com
+
+# File Upload Configuration
+MAX_FILE_SIZE=100MB
+UPLOAD_PATH=./uploads
+ALLOWED_FILE_TYPES=video/mp4,video/avi,video/mov,application/pdf,image/jpeg,image/png
 
 # Security Configuration
 BCRYPT_ROUNDS=12
@@ -19,9 +48,29 @@ RATE_LIMIT_MAX_REQUESTS=100
 AUTH_RATE_LIMIT_MAX=5
 
 # CORS Configuration
-CORS_ORIGIN=http://localhost:3000,http://127.0.0.1:3000
+CORS_ORIGIN=http://localhost:3000,https://melies.herokuapp.com
 CORS_CREDENTIALS=true
 
 # Logging Configuration
 LOG_LEVEL=info
+LOG_FILE_PATH=./logs/
+
+# Monitoring Configuration
+SENTRY_DSN=your_sentry_dsn_for_error_tracking
+ANALYTICS_ID=your_google_analytics_id
+
+# Feature Flags
+ENABLE_REGISTRATION=false
+ENABLE_PASSWORD_RESET=true
+ENABLE_EMAIL_NOTIFICATIONS=true
+ENABLE_FILE_UPLOAD=true
+
+# External Services
+OPENAI_API_KEY=your_openai_api_key_for_ai_assistant
+GOOGLE_CLIENT_ID=your_google_oauth_client_id
+GOOGLE_CLIENT_SECRET=your_google_oauth_client_secret
+
+# Development Only
+DEBUG=melies:*
+MOCK_EXTERNAL_SERVICES=true
