+++ .gitignore
@@ -1,0 +1,9 @@
+# Logs
+logs
+*.log
+
+# Environment
+.env
+
+# Dependencies
+node_modules
